React UI: Remove need to import CSS (#45)
* build: CSS is now included in component import instead of user needing
to import CSS

* fix(ui): add safeguard to ensure dist/tailwind.css exists before injecting

* fix(ui)!: remove the need to import our css file

- users no longer need to import our css file as we are
injecting the css into the head.
- this means that users will automatically have our css without
having to import the css file.

BREAKING CHANGE: We are no longer exporting our css file, instead
we are injecting the css into the head. Users will no longer be able
to import our css file as we are no longer exporting it.

---------

Co-authored-by: Brenden Manquen <bmanquen1993@gmail.com>