fix(core,ui): update the zod schemas to match the updated endpoints
* fix(core,ui): update the zod schemas to match the updated endpoints

- update zod schemas to match the new endpoint responses for
retrieving bible books, chapters, and verses.
- update test mocks to match the new endpont responses for
retrieving bible books, chaptes, and verses.

* fix: update the zod schemas and types for passages apis

* chore(ui,hooks): make the css for sdk more manageable

- wrapped all children of the YouVersionProvider with a div
containing the data-yv-sdk attribute so that all css prefixed
with yv: in our sdk will properly render.
- The one exception to this is anything rendered in a portal such
as popovers will still need to put this attribute on the root of
the portal as these render outside of the YouVersionProvider.